+++ src/main/java/io/blackground/jobfinder/services/CountriesService.java
@@ -21,9 +21,6 @@ public class CountriesService {
 
 	private final CountriesRepository countriesRepository;
 
-	/**
-	 * @param taskRepository
-	 */
 	public CountriesService(CountriesRepository countriesRepository) {
 		super();
 		this.countriesRepository = countriesRepository;
@@ -42,12 +39,12 @@ public class CountriesService {
 		countriesRepository.save(country);
 	}
 
-	public void delete(int id) {
+	public void delete(long id) {
 		countriesRepository.delete(id);
 	}
 
-	public Countries findCountry(int id) {
-		return countriesRepository.findOne(id);
+	public Countries findCountry(long id) {
+		return countriesRepository.findOne((id));
 	}
 
 }
