+++ src/main/java/io/blackground/jobfinder/services/IndustryService.java
@@ -48,7 +48,7 @@ public class IndustryService {
 		industryRepository.delete(id);
 	}
 
-	public Industry findById(int id) {
+	public Industry findById(long id) {
 		return industryRepository.findById(id);
 	}
